requirements.txt
websocket-client
numpy
requests